helperfunctions/fieldelement.py
class FieldElement:
    def __init__(self, num: int, prime: int):
        if num >= prime or num < 1:
            error = f"Num {num} not in field range 1 to {prime-1}"
            raise ValueError(error)
        self.num = num
        self.prime = prime


    def __repr__(self) -> str:
        return f"FieldElement_{self.num}({self.prime})"


    def __eq__(self, other: "FieldElement") -> bool:
        if other is None:
            return False
        return self.num == other.num and self.prime == other.prime


    def __ne__(self, other: "FieldElement") -> bool:
        return not self == other


    def __add__(self, other: "FieldElement") -> "FieldElement":
        if self.prime != other.prime:
            raise TypeError("Numbers must be in the same modulo field")
        num = (self.num + other.num) % self.prime
        return self.__class__(num, self.prime)


    def __sub__(self, other: "FieldElement") -> "FieldElement":
        if self.prime != other.prime:
            raise TypeError("Numbers must be in the same modulo field")
        num = (self.num - other.num) % self.prime
        return self.__class__(num, self.prime)


    def __mul__(self, other: "FieldElement") -> "FieldElement":
        if self.prime != other.prime:
            raise TypeError("Numbers must be in the same modulo field")
        num = (self.num * other.num) % self.prime
        return self.__class__(num, self.prime)


    def __pow__(self, exponent: int) -> "FieldElement":
        n = exponent % (self.prime - 1)
        num = pow(self.num, n, self.prime)
        return self.__class__(num, self.prime)


    def __truediv__(self, other: "FieldElement") -> "FieldElement":
        if self.prime != other.prime:
            raise TypeError("Numbers must be in the same modulo field")
        num = (self.num * pow(other.num, self.prime - 2, self.prime)) % self.prime
        return self.__class__(num, self.prime)


    def __rmul__(self, coefficient: int) -> "FieldElement":
        num = (self.num * coefficient) % self.prime
        return self.__class__(num=num, prime=self.prime)


    def sqrt(self) -> "FieldElement":
        return self ** ((self.prime + 1) // 4)
